rebase trainer
