+++ autogl/module/train/__init__.py
@@ -1,8 +1,14 @@
 import importlib
 import os
-from .base import BaseTrainer, Evaluation, EarlyStopping
 
 TRAINER_DICT = {}
+EVALUATE_DICT = {}
+from .base import (
+    BaseTrainer,
+    Evaluation,
+    BaseNodeClassificationTrainer,
+    BaseGraphClassificationTrainer,
+)
 
 
 def register_trainer(name):
@@ -17,9 +23,6 @@ def register_trainer(name):
         return cls
 
     return register_trainer_cls
-
-
-EVALUATE_DICT = {}
 
 
 def register_evaluate(*name):
@@ -47,14 +50,16 @@ def get_feval(feval):
     raise ValueError("feval argument of type", type(feval), "is not supported!")
 
 
-from .graph_classification import GraphClassificationTrainer
-from .node_classification import NodeClassificationTrainer
+from .graph_classification_full import GraphClassificationFullTrainer
+from .node_classification_full import NodeClassificationFullTrainer
 from .evaluate import Acc, Auc, Logloss
 
 __all__ = [
     "BaseTrainer",
-    "GraphClassificationTrainer",
-    "NodeClassificationTrainer",
+    "BaseNodeClassificationTrainer",
+    "BaseGraphClassificationTrainer",
+    "GraphClassificationFullTrainer",
+    "NodeClassificationFullTrainer",
     "Evaluation",
     "Acc",
     "Auc",
