+++ autogl/module/train/base.py
@@ -1,10 +1,23 @@
 import numpy as np
 from typing import Union, Iterable
-from ..model import BaseModel
+
+import torch
+from ..model import BaseModel, MODEL_DICT
 import pickle
 from ...utils import get_logger
+from . import EVALUATE_DICT
 
 LOGGER_ES = get_logger("early-stopping")
+
+
+def get_feval(feval):
+    if isinstance(feval, str):
+        return EVALUATE_DICT[feval]
+    if isinstance(feval, type) and issubclass(feval, Evaluation):
+        return feval
+    if isinstance(feval, list):
+        return [get_feval(f) for f in feval]
+    raise ValueError("feval argument of type", type(feval), "is not supported!")
 
 
 class EarlyStopping:
@@ -81,17 +94,11 @@ class EarlyStopping:
 class BaseTrainer:
     def __init__(
         self,
-        model: Union[BaseModel, str],
-        optimizer=None,
-        lr=None,
-        max_epoch=None,
-        early_stopping_round=None,
-        device=None,
+        model: BaseModel,
+        device: Union[torch.device, str],
         init=True,
         feval=["acc"],
         loss="nll_loss",
-        *args,
-        **kwargs,
     ):
         """
         The basic trainer.
@@ -103,29 +110,26 @@ class BaseTrainer:
         model: `BaseModel` or `str`
             The (name of) model used to train and predict.
 
-        optimizer: `Optimizer` of `str`
-            The (name of) optimizer used to train and predict.
-
-        lr: `float`
-            The learning rate.
-
-        max_epoch: `int`
-            The max number of epochs in training.
-
-        early_stopping_round: `int`
-            The round of early stop.
-
-        device: `torch.device` or `str`
-            The device where model will be running on.
-
         init: `bool`
             If True(False), the model will (not) be initialized.
-
-        args: Other parameters.
-
-        kwargs: Other parameters.
         """
         super().__init__()
+        self.model = model
+        self.to(device)
+        self.init = init
+        self.feval = get_feval(feval)
+        self.loss = loss
+
+    def to(self, device):
+        """
+        Migrate trainer to new device
+
+        Parameters
+        ----------
+        device: `str` or `torch.device`
+            The device this trainer will use
+        """
+        self.device = torch.device(device)
 
     def initialize(self):
         """Initialize the auto model in trainer."""
@@ -169,8 +173,8 @@ class BaseTrainer:
 
     @classmethod
     def load(cls, path):
-        with open(path, "rb") as input:
-            instance = pickle.load(input)
+        with open(path, "rb") as inputs:
+            instance = pickle.load(inputs)
             return instance
 
     @property
@@ -279,7 +283,21 @@ class BaseTrainer:
 
     def set_feval(self, feval):
         """Set the evaluation metrics."""
-        raise NotImplementedError()
+        self.feval = get_feval(feval)
+
+    def update_parameters(self, **kwargs):
+        """
+        Update parameters of this trainer
+        """
+        for k, v in kwargs.items():
+            if k == "feval":
+                self.set_feval(v)
+            elif k == "device":
+                self.to(v)
+            elif hasattr(self, k):
+                setattr(self, k, v)
+            else:
+                raise KeyError("Cannot set parameter", k, "for trainer", self.__class__)
 
 
 # a static class for evaluating results
@@ -296,7 +314,7 @@ class Evaluation:
         """
         Should return whether this evaluation method is higher better (bool)
         """
-        raise True
+        return True
 
     @staticmethod
     def evaluate(predict, label):
@@ -304,3 +322,84 @@ class Evaluation:
         Should return: the evaluation result (float)
         """
         raise NotImplementedError()
+
+
+class BaseNodeClassificationTrainer(BaseTrainer):
+    def __init__(
+        self,
+        model: Union[BaseModel, str],
+        num_features,
+        num_classes,
+        device="auto",
+        init=True,
+        feval=["acc"],
+        loss="nll_loss",
+    ):
+        self.num_features = num_features
+        self.num_classes = num_classes
+        device = (
+            torch.device("cuda" if torch.cuda.is_available() else "cpu")
+            if device == "auto"
+            else torch.device(device)
+        )
+        if isinstance(model, str):
+            assert model in MODEL_DICT, "Cannot parse model name " + model
+            self.model = MODEL_DICT[model](num_features, num_classes, device, init=init)
+        elif isinstance(model, BaseModel):
+            self.model = model
+        else:
+            raise TypeError(
+                "Model argument only support str or BaseModel, get",
+                type(model),
+                "instead.",
+            )
+        super().__init__(model, device=device, init=init, feval=feval, loss=loss)
+
+    @classmethod
+    def get_task_name(cls):
+        return "GraphClassification"
+
+
+class BaseGraphClassificationTrainer(BaseTrainer):
+    def __init__(
+        self,
+        model: Union[BaseModel, str],
+        num_features,
+        num_classes,
+        num_graph_features=0,
+        device=None,
+        init=True,
+        feval=["acc"],
+        loss="nll_loss",
+    ):
+        self.num_features = num_features
+        self.num_classes = num_classes
+        self.num_graph_features = num_graph_features
+        device = (
+            torch.device("cuda" if torch.cuda.is_available() else "cpu")
+            if device == "auto"
+            else torch.device(device)
+        )
+        if isinstance(model, str):
+            assert model in MODEL_DICT, "Cannot parse model name " + model
+            self.model = MODEL_DICT[model](
+                num_features,
+                num_classes,
+                device,
+                init=init,
+                num_graph_features=num_graph_features,
+            )
+        elif isinstance(model, BaseModel):
+            self.model = model
+        else:
+            raise TypeError(
+                "Model argument only support str or BaseModel, get",
+                type(model),
+                "instead.",
+            )
+
+        super().__init__(model, device=device, init=init, feval=feval, loss=loss)
+
+    @classmethod
+    def get_task_name(cls):
+        return "NodeClassification"
